requirements.txt
Django==2.2.0
psycopg2-binary>=2.8
djangorestframework==3.11.0
django-cors-headers==3.2.1
numpy==1.18.1
pandas==1.0.0
nltk==3.4.5
Keras==2.3.1
tensorflow==1.15.0
scikit-learn==0.23.1
xlrd >= 1.0.0
mysqlclient==1.4.6
